refactor: #59 로그아웃 API 리펙토링
로그아웃 API가 호출될 경우, 만약 해당 사용자가 로그인 중이 아니라면 예외를 발생하도록 리펙토링하였다.

- AuthExceptionCode에 관련 예외 코드 추가
- 로그인 중인 사용자가 없을경우 발생하는 사용자 정의 예외 추가 및 핸들러에 등록
- 토큰에서 조회한 사용자의 personalId를 기반으로 해당 사용자의 refresh token이 존재하는지 조회하는 쿼리 메서드 생성
- JwtService에 관련 로직 추가

+++ spadeworker/src/main/java/site/devtown/spadeworker/domain/auth/service/JwtService.java
@@ -9,6 +9,7 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import site.devtown.spadeworker.domain.auth.exception.InvalidTokenException;
+import site.devtown.spadeworker.domain.auth.exception.LoggedUserNotFoundException;
 import site.devtown.spadeworker.domain.auth.model.UserPrincipal;
 import site.devtown.spadeworker.domain.auth.repository.UserRefreshTokenRepository;
 import site.devtown.spadeworker.domain.auth.token.AuthToken;
@@ -108,6 +109,11 @@ public class JwtService {
         AuthToken accessAuthToken = createAuthTokenFromAccessTokenValue(accessToken);
         String userPersonalId = accessAuthToken.getTokenClaims().getSubject();
 
+        // 현재 사용자가 로그인 중이지 않다면 예외 출력
+        if (!isUserLoggedIn(userPersonalId)) {
+            throw new LoggedUserNotFoundException();
+        }
+
         // refresh token 삭제
         userRefreshTokenRepository.deleteAllByPersonalId(userPersonalId);
     }
@@ -157,5 +163,10 @@ public class JwtService {
         return Arrays.stream(userRoles.split(","))
                 .map(SimpleGrantedAuthority::new)
                 .collect(Collectors.toList());
+    }
+
+    // 현재 사용자가 로그인 중인지 체크하는 메소드
+    private boolean isUserLoggedIn(String userPersonalId) {
+        return userRefreshTokenRepository.existsByPersonalId(userPersonalId);
     }
 }
